# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.ApproveValuationRequest:
    properties:
      comment:
        example: Valuation approved after review
        maxLength: 500
        type: string
    type: object
...
  handlers.CreateValuationRequest:
    properties:
      district:
        type: string
      include_comparables:
        type: boolean
      include_factors:
        type: boolean
      property_size_sqm:
        type: number
      property_type:
        enum:
        - residential
        - commercial
        - agricultural
        - industrial
        type: string
      sector:
        type: string
    required:
    - district
    - property_size_sqm
    - property_type
    type: object
...
  handlers.PropertySummary:
    properties:
      district:
        type: string
      id:
        type: integer
      land_size:
        type: number
      price:
        type: number
      property_type:
        type: string
      sector:
        type: string
      title:
        type: string
    type: object
  handlers.RejectValuationRequest:
    properties:
      comment:
        example: Additional details on rejection
        maxLength: 500
        type: string
      reason:
        example: Incorrect property type classification
        maxLength: 500
        type: string
    required:
    - reason
    type: object
...
  handlers.ValuationResponse:
    properties:
      adjustments:
        additionalProperties:
          format: float64
          type: number
        type: object
      base_price:
        type: number
      comparables:
        items:
          $ref: '#/definitions/handlers.PropertySummary'
        type: array
      confidence:
        type: number
      confidence_level:
        type: string
      estimated_value_rwf:
        type: number
      factors:
        items:
          $ref: '#/definitions/models.ValuationFactor'
        type: array
      gazette_reference:
        type: string
      recommendations:
        items:
          type: string
        type: array
      updated_at:
        type: string
      zone_coefficient:
        type: number
    type: object
...
  models.ValuationFactor:
    properties:
      description:
        type: string
      impact:
        description: positive, negative, neutral
        type: string
      name:
        type: string
      value:
        type: number
      weight:
        type: number
    type: object
  utils.APIError:
    properties:
      details:
        type: string
      message:
        type: string
    type: object
  utils.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/utils.APIError'
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: api.landvaluationsystem.rw
info:
  contact:
    email: support@landvaluationsystem.rw
    name: API Support
    url: http://www.landvaluationsystem.rw/support
  description: Rwanda Land Valuation System with Official Gazette Integration
  license:
    name: Proprietary
    url: http://www.landvaluationsystem.rw/license
  termsOfService: http://swagger.io/terms/
  title: Land Valuation System API
  version: "1.0"
paths:
...
  /admin/valuations/{id}/approve:
    post:
      consumes:
      - application/json
      description: Admin endpoint to approve a property valuation
      parameters:
      - description: Valuation ID
        in: path
        name: id
        required: true
        type: string
      - description: Approval details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ApproveValuationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Approve a valuation (Admin)
      tags:
      - admin
      - valuations
  /admin/valuations/{id}/reject:
    post:
      consumes:
      - application/json
      description: Admin endpoint to reject a property valuation with reason
      parameters:
      - description: Valuation ID
        in: path
        name: id
        required: true
        type: string
      - description: Rejection details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RejectValuationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Reject a valuation (Admin)
      tags:
      - admin
      - valuations
...
  /api/v1/valuations:
    get:
      description: Get a list of all valuations created by the authenticated user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handlers.ValuationResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: List user's valuations
      tags:
      - valuations
    post:
      consumes:
      - application/json
      description: Calculate estimated property value based on location, size, and
        type
      parameters:
      - description: Valuation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateValuationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.ValuationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Calculate property valuation
      tags:
      - valuations
  /api/v1/valuations/{id}:
    get:
      description: Retrieve a specific valuation result by ID
      parameters:
      - description: Valuation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.ValuationResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Get valuation by ID
      tags:
      - valuations
  /api/v1/valuations/history:
    get:
      description: Retrieve all valuations performed on a specific property
      parameters:
      - description: Property ID
        in: query
        name: property_id
        required: true
        type: string
      - default: 10
        description: Maximum number of records
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handlers.ValuationResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get valuation history for a property
      tags:
      - valuations
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"